+++ app/src/main/java/io/apicurio/registry/rest/v3/GroupsResourceImpl.java
@@ -181,6 +181,11 @@ public class GroupsResourceImpl extends AbstractResourceImpl implements GroupsRe
     @Audited(extractParameters = { "0", KEY_GROUP_ID, "1", KEY_ARTIFACT_ID })
     @Authorized(style = AuthorizedStyle.GroupAndArtifact, level = AuthorizedLevel.Write)
     public void deleteArtifact(String groupId, String artifactId) {
+        if (!restConfig.isArtifactDeletionEnabled()) {
+            throw new NotAllowedException("Artifact deletion operation is not enabled.", HttpMethod.GET,
+                    (String[]) null);
+        }
+
         requireParameter("groupId", groupId);
         requireParameter("artifactId", artifactId);
 
@@ -240,6 +245,11 @@ public class GroupsResourceImpl extends AbstractResourceImpl implements GroupsRe
     @Override
     @Authorized(style = AuthorizedStyle.GroupOnly, level = AuthorizedLevel.Write)
     public void deleteGroupById(String groupId) {
+        if (!restConfig.isGroupDeletionEnabled()) {
+            throw new NotAllowedException("Group deletion operation is not enabled.", HttpMethod.GET,
+                    (String[]) null);
+        }
+
         storage.deleteGroup(groupId);
     }
 
@@ -631,6 +641,11 @@ public class GroupsResourceImpl extends AbstractResourceImpl implements GroupsRe
     @Audited(extractParameters = { "0", KEY_GROUP_ID })
     @Authorized(style = AuthorizedStyle.GroupOnly, level = AuthorizedLevel.Write)
     public void deleteArtifactsInGroup(String groupId) {
+        if (!restConfig.isArtifactDeletionEnabled()) {
+            throw new NotAllowedException("Artifact deletion operation is not enabled.", HttpMethod.GET,
+                    (String[]) null);
+        }
+
         requireParameter("groupId", groupId);
 
         storage.deleteArtifacts(new GroupId(groupId).getRawGroupIdWithNull());
